Refs #39 Update table automatically when expense is added or removed
